app/requirements.txt
requests
flask-marshmallow